feat: add fastify adapter support & make minimal code to run nest on both adapters
