--- dev/nest-node/.prettierrc
@@ -1,4 +1,0 @@
-{
-  "singleQuote": true,
-  "trailingComma": "all"
-}
