--- dev/nest-node/nest-cli.json
@@ -1,7 +1,0 @@
-{
-  "$schema": "https://json.schemastore.org/nest-cli",
-  "collection": "@nestjs/schematics",
-  "sourceRoot": ".",
-  "entryFile": "index",
-  "compilerOptions": { "deleteOutDir": true }
-}
