+++ dev/nest-node/package.json
@@ -1,39 +1,25 @@
 {
-  "name": "nest-typescript-starter",
+  "name": "nest-multi-platform",
   "version": "1.0.0",
   "private": true,
-  "description": "Nest TypeScript starter repository",
   "license": "MIT",
   "scripts": {
-    "build": "nest build",
-    "dev": "nest start --watch",
-    "format": "prettier --write \"./**/*.{ts,js}\"",
-    "lint": "eslint \"./**/*.{ts,js}\" --fix",
-    "start": "nest start"
+    "build": "bun build src/nest-*.ts --outdir ../../src/node --packages external --target node",
+    "dev": "bun src/nest-express.ts"
   },
   "dependencies": {
     "@nestjs/common": "^10.4.8",
     "@nestjs/core": "^10.4.8",
     "@nestjs/platform-express": "^10.4.8",
+    "@nestjs/platform-fastify": "10.4.8",
     "reflect-metadata": "^0.2.2",
     "rxjs": "^7.8.1"
   },
   "devDependencies": {
-    "@nestjs/cli": "^10.4.7",
-    "@nestjs/schematics": "^10.2.3",
-    "@types/express": "^4.17.17",
+    "@types/express": "^4.17.21",
     "@types/node": "^22.9.0",
-    "@typescript-eslint/eslint-plugin": "^8.15.0",
-    "@typescript-eslint/parser": "^8.15.0",
-    "eslint": "^8.42.0",
-    "eslint-config-prettier": "^9.1.0",
-    "eslint-plugin-prettier": "^5.2.1",
-    "prettier": "^3.3.3",
-    "source-map-support": "^0.5.21",
-    "ts-loader": "^9.5.1",
-    "ts-node": "^10.9.2",
-    "tsconfig-paths": "^4.2.0",
+    "fastify": "4.28.1",
     "typescript": "^5.6.3"
   },
-  "packageManager": "pnpm@9.13.2"
+  "packageManager": "bun@1.1.34"
 }
